Rename models to model under domain layer
