+++ Abhi-AnimeX/app/src/main/java/com/abhijith/animex/ui/screens/animedetails/AnimeDetails.kt
@@ -40,7 +40,7 @@ import androidx.lifecycle.viewmodel.compose.viewModel
 import coil.compose.AsyncImage
 import coil.request.ImageRequest
 import com.abhijith.animex.R
-import com.abhijith.animex.domain.models.AnimeItem
+import com.abhijith.animex.domain.model.AnimeItem
 import com.abhijith.animex.ui.components.GenreTag
 import com.abhijith.animex.ui.components.ScoreTag
 import com.abhijith.animex.ui.components.StatItem
