+++ Abhi-AnimeX/app/src/main/java/com/abhijith/animex/ui/screens/animelist/AnimeListUiState.kt
@@ -1,6 +1,6 @@
 package com.abhijith.animex.ui.screens.animelist
 
-import com.abhijith.animex.domain.models.AnimeItem
+import com.abhijith.animex.domain.model.AnimeItem
 
 sealed class AnimeListUiState {
     data object Loading : AnimeListUiState()
